Conflicts:
	src/main/java/org/olat/repository/ui/catalog/CatalogNodeController.java
	src/main/java/org/olat/upgrade/_spring/databaseUpgradeContext.xml

+++ src/main/resources/database/mysql/alter_15_pre_0_to_15_pre_4.sql
@@ -43,7 +43,7 @@ create table o_grad_time_record (
    id bigint not null auto_increment,
    creationdate datetime not null,
    lastmodified datetime not null,
-   g_time int8 default 0 not null,
+   g_time bigint default 0 not null,
    g_date_record date not null,
    fk_assignment bigint,
    fk_grader bigint not null,
